% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Albaniadata.R
\docType{data}
\name{Albania2013}
\alias{Albania2013}
\title{Albania data}
\format{csv, shp, dbf, shx, prj}
\description{
Electoral data: Albania2013.csv (parliament); Geodata: Albania2013_communes, Albania2013_prefectures (.shp, .dbf, .shx, .prj)
}
\keyword{Albania}
\keyword{data}
