% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BasicElectionForensics.R
\name{BasicElectionForensics}
\alias{BasicElectionForensics}
\title{Basic Election Forensics function}
\usage{
BasicElectionForensics(data, Candidates, Level,
                             TotalReg, TotalVotes, Methods, R=1000, cores=2)
}
\arguments{
\item{data}{data frame}

\item{Candidates}{variable name referring to vote counts for candidates/parties}

\item{Level}{variable name depicting the level of analysis (whole dataset by default)}

\item{TotalReg}{variable name for the total number of eligible voters}

\item{TotalVotes}{variable name for the total number of ballots cast}

\item{Methods}{\itemize{
  \item "_2BL" - second-digit mean test (to be compared to the mean value Benford's Law)
  \item "LastC" - last-digit mean test (to be compared to the mean value implied by uniformly distributed last digits)
  \item "P05s" - percentage last-digit 0/5 indicator mean (to be compared to the mean value implied by uniformly distributed percentages)
  \item "C05s" - count last-digit 0/5 indicator mean count (last-digit 0/5 indicator mean)
  \item "Sobyanin" - Sobynin-Sukhovolsky measure
  \item "Skew" - skewness (the extent to which a variable departs from a normal distribution by being asymmetric)
  \item "Kurt" - Kurt (the extent to which a variable departs from a normal distribution by being spread out too much or not enough)
  \item "DipT" - Unimodality test (tests whether the distribution of a variable departs from unimodality)
  \item "Correlation" - correlation coefficient between turnout and vote share
}}

\item{R}{number of simulations (1000  simulations by default)}

\item{cores}{number of cores for parallel computing (2 cores by default)}
}
\value{
A list with the following parameters:
\itemize{
  \item table - dataframe with results
  \item html -  html table with results
  \item tex - tex table with results
  \item sigMatrix - significance matrix
}
}
\description{
This function helps to run basic election forensics methods: Benford's 2nd Digit, Last Digit (Counts),
Counts (05s), Multimodality Test, Sobyanin-Sukhovolsky test and Correlation test.
}
\examples{
library(EFToolkit)

dat<-read.csv(system.file("Albania2013.csv", package="EFToolkit"))

#NB! R=100 to speed up computations for this example.
eldata<-BasicElectionForensics(dat,
                          Candidates=c("C035", "C050"),
                          Level="Prefectures", TotalReg="Registered",
                          TotalVotes="Ballots",
                          Methods=c("P05s", "C05s", "_2BL", "Sobyanin",
                         "DipT", "Skew", "Kurt", "Correlation"), cores=2, R=100)
eldata

}
