% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BuildMap.R
\name{BuildMap}
\alias{BuildMap}
\title{Build Map function}
\usage{
BuildMap(eforensicsdata, geodata, Geoindex, Colorsig, xlab)
}
\arguments{
\item{eforensicsdata}{BasicElectionForensics() object}

\item{geodata}{sf object}

\item{Geoindex}{name of index variable from sf object used to merge geodata with BasicElectionForensics() object}

\item{Colorsig}{only statistically significant estimates are mapped (FALSE by default)}

\item{xlab}{X axis label}
}
\value{
A list containing the following parameters:
\itemize{
  \item figures - list of constructed figures
  \item Colorsig -  external parameter
  \item shpdata - updated sf object
  \item creationdate - date/time of analysis
}
}
\description{
This function constructs the map based on the results from BasicElectionForensics() object.
}
\examples{
library(EFToolkit)
library(sf)

dat<-read.csv(system.file("extdata/Albania2013.csv", package="EFToolkit"))
eldata<-BasicElectionForensics(dat,
                              Candidates=c("C035", "C050"),
                              Level="Prefectures", TotalReg="Registered",
                              TotalVotes="Ballots",
                              Methods=c("P05s", "C05s"), R=100)
geodata<-st_read(system.file("extdata/Albania2013_prefectures.shp",
                          package="EFToolkit"), quiet = TRUE)

figures<-BuildMap(eforensicsdata=eldata, geodata=geodata, Geoindex="Level")
}
