% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ClusterAnalysis.R
\name{ClusterAnalysis}
\alias{ClusterAnalysis}
\title{Cluster Analysis}
\usage{
ClusterAnalysis(geodata, Vars, IndexCL, cores)
}
\arguments{
\item{geodata}{shapefile object within a list (the algorithm also takes both polygon and point shapefiles within a list) or BuildMap object().}

\item{Vars}{variable names used for geographic clustering tests}

\item{IndexCL}{index variable name used for merging polygon and point shapefiles}

\item{cores}{number of cores for parallel computing (2 cores by default)}
}
\value{
A list of maps with results from cluster analysis:
\itemize{
  \item table - dataframe with results
  \item html -  html table with results
  \item tex - tex table with results
  \item sigMatrix - significance matrix
}
}
\description{
This function implements geographic clustering tests: Getis-Ord Gi and local Moran's I.
}
\examples{
library(EFToolkit)
library(rgdal)

dat<-read.csv(system.file("Albania2013.csv", package="EFToolkit"))
#Obtain election forensics estimates
eldata<-BasicElectionForensics(dat,
                              Candidates=c("C035", "C050"),
                              Level="Prefectures", TotalReg="Registered",
                              TotalVotes="Ballots",
                              Methods=c("P05s", "C05s"), R=100)

#Create the map with results
geodata<-readOGR(system.file("Albania2013_prefectures.shp", package="EFToolkit"), verbose = FALSE)
figures<-BuildMap(eforensicsdata=eldata, geodata=geodata, Geoindex="Level", Colorsig=FALSE)

#Using the mapped results, implement cluster analysis
figureC<-ClusterAnalysis(figures, Vars=c("C050_P05s", "C050_C05s"))


}
