% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ComputeFiniteMixtureModel.R
\name{ComputeFiniteMixtureModel}
\alias{ComputeFiniteMixtureModel}
\title{Finite Mixture Model}
\usage{
ComputeFiniteMixtureModel(dat, MainCandidate="Votes", TotalReg="NVoters",
                                 TotalVotes="NValid", cores=2, itstartmax=1)
}
\arguments{
\item{dat}{electoral data.}

\item{MainCandidate}{variable name for major/incumbent candidate}

\item{TotalReg}{variable name for the total number of eligible voters}

\item{TotalVotes}{variable name for the total number of ballots cast}

\item{cores}{number of cores for parallel computing (2 cores by default)}

\item{itstartmax}{maximum number of iterations}
}
\value{
A list of FMM estimates.
}
\description{
This function implements the Finite Mixture Model developed and coded by Walter Mebane.
}
\examples{
library(EFToolkit)

#Russia examples
dat<-read.csv(system.file("extdata/ruspres2000.csv", package="EFToolkit"))
dat<-subset(dat, select=c("regname", "NVoters", "NValid", "Votes"))
datc<-dat[dat$region=="Volgogradskaya Oblast`",]
#Do not run: takes long time to compute
\dontrun{
res <- ComputeFiniteMixtureModel(datc,
       MainCandidate = "Votes",
       TotalReg = "NVoters",
       TotalVotes = "NValid")
 }
}
