% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ComputeKlimekModel.R
\name{ComputeKlimekModel}
\alias{ComputeKlimekModel}
\title{Compute Klimek Model function}
\usage{
ComputeKlimekModel(data, Candidates, Level, TotalReg, TotalVotes, R=1000, cores=2)
}
\arguments{
\item{data}{data frame}

\item{Candidates}{variable name referring to vote counts for candidates/parties}

\item{Level}{variable name depicting the level of analysis (whole dataset by default)}

\item{TotalReg}{variable name for the total number of eligible voters}

\item{TotalVotes}{variable name for the total number of ballots cast}

\item{R}{number of simulations (1000  simulations by default)}

\item{cores}{number of cores for parallel computing (2 cores by default)}
}
\value{
A list with the following parameters:
\itemize{
  \item table - data frame with results
  \item html -  html table with results
  \item tex - tex table with results
  \item sigMatrix - significance matrix
}
}
\description{
This function helps to estimate the Klimek et al.(2012) model.
}
\examples{
library(EFToolkit)

dat<-read.csv(system.file("Albania2013.csv", package="EFToolkit"))

#NB! R=100 to speed up computations for this example.
klimek<-ComputeKlimekModel(dat, Candidates="C050", Level="National",
                        TotalReg="Registered", TotalVotes="Ballots", cores=1, R=100)

}
