% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ComputeShpilkinMethod.R
\name{ComputeShpilkinMethod}
\alias{ComputeShpilkinMethod}
\title{Shpilkin's Method}
\usage{
ComputeShpilkinMethod(data, Candidates, CandidatesText = NULL, MainCandidate, TotalReg, TotalVotes=NULL,
                             Level = NULL, Methodmax = "M1", WindowSize = 5, MaxtThreshold = 0.8, FigureName,  Colors = NULL)
}
\arguments{
\item{data}{electoral data.}

\item{Candidates}{vector of variable names for all candidates/parties participated in the election}

\item{CandidatesText}{vector of candidates/parties' names participated in the election used to draw the figure}

\item{MainCandidate}{variable name for main/incumbent candidate}

\item{TotalReg}{variable name for the total number of eligible voters}

\item{TotalVotes}{variable name for the total number of ballots cast (if ommited, computed based on the votes for Candidates)}

\item{Level}{variable name depicting the level of analysis ("National", i.e. whole dataset by default)}

\item{Methodmax}{clean peak search
\itemize{
  \item M0 - absolute clean peak search on the left handside from official turnout
  \item M1 - absolute clean peak search on the left handside from official turnout with k-means clustering
  \item M2 - relative search on the left handside from official turnout within a range defined by WindowSize
}}

\item{WindowSize}{define WindowSize for M0 and M1.  Algorithm searches for max value change within prespecified WindowSize (by default WindowSize=5\%)}

\item{MaxtThreshold}{anomalous turnout threshold (by default 0.8)}

\item{FigureName}{figure's name}

\item{Colors}{vector of colors per each candidate/party (Colors=NULL, i.e. randomly generated by default)}
}
\value{
list containing results of analysis
\itemize{
  \item list_graphs - list of graphs
  \item stats_summary -  table with results for analysis of the whole dataset (for Level!=NULL: data are also
                         summed across the units, aggregation error is computed)
  \item Level -  external parameter
  \item creationdate - date/time of analysis
  list_graphs = list_graphs, stats_summary = stats_table, stats_level=stats_level
}
}
\description{
This function implements the revised version of Shpilkin's method (NB! It doesn't replicate original method fully).
}
\examples{
library(EFToolkit)

dat<-read.csv(system.file("ruspres2018.csv", package="EFToolkit"))
res<-ComputeShpilkinMethod(dat, Candidates=c("P1", "P2",  "P3",  "P4", "P5", "P6", "P7", "P8" ),
                    CandidatesText=c("Baburin", "Grudinin",  "Zhirinovsky",  "Putin", "Sobchak",
                             "Suraikin", "Titov", "Yavlinsky"),
                    MainCandidate="P4",
                    TotalReg="NVoters",
                    TotalVotes=c('C9', 'C10'),
                    Methodmax="M1",
                    FigureName="Russian Presidential Elections, 2018",
                    Level="region",
                    MaxtThreshold=0.85)
}
