% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{internal-functions}
\alias{internal-functions}
\alias{var_addition}
\title{Internal functions for election fraud detection}
\usage{
var_addition(
  data,
  Level,
  MainCandidate,
  Candidates,
  TotalVotes = NULL,
  CandidatesText = NULL,
  TotalReg
)
}
\value{
Returns various data outputs.
}
\description{
Various functions for detecting election anomalies using statistical methods
}
\keyword{internal}
