% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Russiadata.R
\docType{data}
\name{ruspres2018}
\alias{ruspres2018}
\title{Russian electoral data}
\format{
csv
}
\description{
Russia electoral data collected from 2018 Russian presidential elections
}
\keyword{Russian}
\keyword{elections}
