% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ComputeShpilkinMethod.R
\name{ComputeShpilkinMethod}
\alias{ComputeShpilkinMethod}
\title{Shpilkin's Method}
\usage{
ComputeShpilkinMethod(data, Candidates, CandidatesText=NULL, MainCandidate, TotalReg, Level=NULL,
                             Methodmax="M1", MaxtThreshold=0.8, WindowSize=5, FigureName)
}
\arguments{
\item{data}{electoral data.}

\item{Candidates}{vector of variable names for all candidates/parties participated in the election}

\item{CandidatesText}{vector of candidates/parties' names participated in the election used to draw the figure}

\item{MainCandidate}{variable name for main/incumbent candidate}

\item{TotalReg}{variable name for the total number of eligible voters}

\item{Level}{run algorithm for administrative subdivisions}

\item{Methodmax}{clean peak search
\itemize{
  \item M1 - relative search on the left handside from official turnout within a range defined by WindowSize
  \item M2 - absolute clean peak search on the left handside from official turnout
}}

\item{MaxtThreshold}{anomalous turnout threshold (by default 0.8)}

\item{WindowSize}{define WindowSize for M1.  Algorithm searches for max change within prespecified WindowSize (by default WindowSize=5\%)}

\item{FigureName}{figure's name}
}
\value{
list containing results of analysis
\itemize{
  \item list_graphs - list of graphs
  \item stats_table -  table with results
  \item Level -  external parameter
  \item creationdate - date/time of analysis
}
}
\description{
This function implements the revised version of Shpilkin's method (NB! It doesn't replicate original method fully).
}
\examples{
library(EFToolkit)

dat<-read.csv(system.file("ruspres2018.csv", package="EFToolkit"))
res<-ComputeShpilkinMethod(dat, Candidates=c("P1", "P2",  "P3",  "P4", "P5", "P6", "P7", "P8" ),
                    CandidatesText=c("Baburin", "Grudinin",  "Zhirinovsky",  "Putin", "Sobchak",
                             "Suraikin", "Titov", "Yavlinsky"),
                    MainCandidate="P4",
                    TotalReg="NVoters",
                    Methodmax="M2",
                    FigureName="Russian Presindential Elections, 2018",
                    Level="region",
                    MaxtThreshold=0.85)
}
